First real Angular implementation
Created a new app-folder for all angular-app related files.
app/app.js is the central of it all, and it uses currently only
projectControllers-module which has projectListCtrl -controller.
app.js creates the root app, and defines routes, as demonstrated by
links in index.html currently.

projectList.html -view uses the projectListCtrl -controller to show
simple list of projects, which are currently defined for testing
purposes in api/testdata/projects.json (until we get the actual api
running).

I also created projectServices, but didn't get it to work. Need to look
closer to it later, reference for using services:
https://docs.angularjs.org/tutorial/step_11

HomeControllers is also defined, but not used currently, implementation
is just a stub currently.