Add a rules variable and method to Model
The rules variable is an array with key => value pairs, where key is
the name of a class variable and value is the validation method used
to validate it. The validation methods are static methods in Validator.

The setRules method is used to populate the rules variable on object
construction. By doing the population the rules variable can be used
later to set and get object variables, for example when reading post
data or when validating object attributes.