Generalize loadPost from User, move it to Model
The loadPost method is used to get post data and populate object
attributes with it. The implementation in User was done by defining
all attributes by hand. The method is made more useful by generalizing
it to use the new rules variable array, and moving it to Model so
other child classes can also benefit from it.